Add minimal support for 'tin' and force-adding of subreddits
The 'tin' newsreader requires the server to advertise READER mode.  It also
can't select groups it hasn't previously seen in the LIST ACTIVE response,
even if you type in their names with the 'subscribe pattern' command.  For
this reason this commit adds a command line option to add new subreddits to
the database.  It only needs doing once.